Start work on CoffeeScript Rewrite
I may or may not take up maintanence of 4chan SS, which would
involve bringing it up to date with the latest 4chan X version(s)
and 4chan Markup. But before I do that, since there is no current
maintainer, I want to bring it up to my style of coding and, as an
exercise in code comprehension, rewrite the entire thing in
CoffeeScript. How successful I am in this effort will help me to
determine whether or not I will actually maintain this or not.

Obviously, there's no promises here, and this is the first time
I've taken to contributing to 4chan SS at all. But if everything
goes well, I'm sure someone will appreciate these efforts. I'd hate
to see the time and effort put into 4chan SS go to waste.